MLZoomCamp - Capstone/requirements.txt
flask
grpcio
tensorflow-serving-api
tensorflow
Pillow
numpy

